stream/predictor/requirements.txt
numpy
faust-streaming